requirements.txt
click==7.1.2
dataclasses-json==0.5.1
Flask==1.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.13.0
six==1.15.0
stringcase==1.2.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
Werkzeug==1.0.1
Inject==4.2.0
pysqlite3==0.4.3
